requirements.txt
attrs==25.3.0
boto3==1.40.10
botocore==1.40.10
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
execnet==2.1.1
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.0
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.1
pytest-check==2.5.3
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.1
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
